refactor: extracts profile update mutation hook to generic location
