+++ src/v2/Components/Inquiry/__tests__/Views/InquiryBasicInfo.jest.tsx
@@ -41,6 +41,7 @@ describe("InquiryBasicInfo", () => {
     ;(useInquiryContext as jest.Mock).mockImplementation(() => ({
       next: mockNext,
       setContext: () => {},
+      relayEnvironment: { current: null },
     }))
   })
